+++ src/ForeverFactory/Customizations/CustomizeFactoryOptions.cs
@@ -1,62 +1,15 @@
 ﻿using System;
-using System.Collections.Generic;
 using ForeverFactory.Behaviors;
-using ForeverFactory.Generators;
 using ForeverFactory.Generators.Transforms;
 
 namespace ForeverFactory.Customizations
 {
-    internal class OptionsCollector<T> : IOptionsCollector<T>
-        where T : class
-    {
-        private readonly Action<ICustomizeFactoryOptions<T>> _customize;
-        private readonly ObjectFactoryOptions<T> _options;
-
-        public OptionsCollector(Action<ICustomizeFactoryOptions<T>> customize)
-        {
-            _customize = customize;
-            _options = new ObjectFactoryOptions<T>();
-        }
-
-        public IObjectFactoryOptions<T> Collect()
-        {
-            var customizationOptions = new CustomizeFactoryOptions2<T>(_options);
-            _customize.Invoke(customizationOptions);
-
-            return _options;
-        }
-        
-        internal void UpdateConstructor(Func<T> customConstructor)
-        {
-            _options.CustomConstructor = customConstructor;
-        }
-        
-        internal void UpdateBehavior(Behavior behavior)
-        {
-            _options.SelectedBehavior = behavior;
-        }
-    }
-
-    internal class ObjectFactoryOptions<T> : IObjectFactoryOptions<T>
-        where T : class
-    {
-        public Func<T> CustomConstructor { get; internal set; }
-        public Behavior SelectedBehavior { get; internal set; }
-        public IList<Transform<T>> Transforms { get; }
-
-        public ObjectFactoryOptions()
-        {
-            Transforms = new List<Transform<T>>();
-            SelectedBehavior = new DoNotFillBehavior();
-        }
-    }
-
-    internal class CustomizeFactoryOptions2<T> : ICustomizeFactoryOptions<T>
+    internal class CustomizeFactoryOptions<T> : ICustomizeFactoryOptions<T>
         where T : class
     {
         private readonly ObjectFactoryOptions<T> _objectFactoryOptions;
 
-        public CustomizeFactoryOptions2(ObjectFactoryOptions<T> objectFactoryOptions)
+        public CustomizeFactoryOptions(ObjectFactoryOptions<T> objectFactoryOptions)
         {
             _objectFactoryOptions = objectFactoryOptions;
         }
