+++ tests/ForeverFactory.Tests/Generators/ObjectFactoryTests.cs
@@ -1,7 +1,6 @@
 ﻿using FluentAssertions;
 using ForeverFactory.Customizations;
 using ForeverFactory.Generators;
-using ForeverFactory.Generators.Transforms;
 using ForeverFactory.Tests.Factories.CustomizedFactories.ExampleFactories;
 using Xunit;
 
@@ -52,25 +51,6 @@ namespace ForeverFactory.Tests.Generators
             var customizeFactoryOptions = new OptionsCollector<Person>(customization =>
             {
                 customization.Set(x => x.FirstName = "Clark");
-            });
-
-            var factory = new ObjectFactory<Person>(customizeFactoryOptions);
-            factory.AddNode(new GeneratorNode<Person>(1));
-            factory.AddNode(new GeneratorNode<Person>(2));
-
-            var persons = factory.Build();
-
-            foreach (var person in persons) 
-                person.FirstName.Should().Be("Clark");
-        }
-
-        // TODO refactor
-        [Fact]
-        public void It_should_apply_default_transforms_to_all_generator_nodes2()
-        {
-            var customizeFactoryOptions = new OptionsCollector<Person>(initialization =>
-            {
-                initialization.Set(x => x.FirstName = "Clark");
             });
 
             var factory = new ObjectFactory<Person>(customizeFactoryOptions);
